resolve cane violations, rename pregnancies helper to patients helper
