+++ app/models/patient.rb
@@ -1,3 +1,4 @@
+# Object representing core patient information and demographic data
 class Patient
   include Mongoid::Document
   include Mongoid::Timestamps
