+++ app/models/pregnancy.rb
@@ -1,3 +1,4 @@
+# Object representing pregnancy outcomes and status of a patient
 class Pregnancy
   include Mongoid::Document
   include Mongoid::Timestamps
